refactor: Decompose bacpacman.py into a package
